pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "krita-webtoon-export"
version = "0.1.0"
description = "Krita-friendly Webtoon export pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.0.0",
  "pytest>=8.0.0",
  "ruff>=0.3.0",
]

[project.scripts]
webtoon-export = "krita_webtoon_export.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
